+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -509,9 +509,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-08-24T09:28:41.056Z"
+    created: Optional[str] = "2026-08-24T14:41:02.470Z"
 
-    updated: Optional[str] = "2026-08-24T09:28:41.056Z"
+    updated: Optional[str] = "2026-08-24T14:41:02.470Z"
 
     updated_by_id: OptionalNullable[str] = UNSET