+++ packages/orq-rc/RELEASES.md
@@ -10409,3 +10409,13 @@ Based on:
 - [python v4.14.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v4.14.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.47 - packages/orq-rc
+
+## 2026-08-24 14:43:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.2 (2.932.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.14.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v4.14.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.48 - packages/orq-rc